@@ -1,22 +1,14 @@
 package mohak.uberux;
 
-import android.app.ActivityOptions;
 import android.content.Context;
 import android.content.Intent;
 import android.graphics.Color;
 import android.os.Handler;
-import android.support.v4.app.ActivityOptionsCompat;
-import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.transition.ChangeBounds;
 import android.transition.Slide;
 import android.transition.Transition;
-import android.transition.TransitionInflater;
-import android.transition.Visibility;
-import android.view.Gravity;
 import android.view.View;
-import android.view.WindowManager;
-import android.view.animation.AccelerateInterpolator;
 import android.view.animation.Animation;
 import android.view.animation.AnimationUtils;
 import android.view.animation.DecelerateInterpolator;
@@ -26,8 +18,9 @@ import android.widget.FrameLayout;
 import android.widget.ImageView;
 import android.widget.LinearLayout;
 import android.widget.TextView;
-import android.widget.Toast;
 
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.core.app.ActivityOptionsCompat;
 
 import com.github.jorgecastilloprz.FABProgressCircle;
 
@@ -36,10 +29,8 @@ import butterknife.ButterKnife;
 import butterknife.OnClick;
 
 import static android.view.Gravity.LEFT;
-import static android.view.Gravity.RIGHT;
 
 public class LoginWithPhone extends AppCompatActivity {
-
     @BindView(R.id.ivback)
     ImageView ivBack;
 
@@ -70,11 +61,9 @@ public class LoginWithPhone extends AppCompatActivity {
         setContentView(R.layout.activity_login_with_phone);
         ButterKnife.bind(this);
         setupWindowAnimations();
-
     }
 
     private void setupWindowAnimations() {
-
         ChangeBounds enterTransition = new ChangeBounds();
         enterTransition.setDuration(1000);
         enterTransition.setInterpolator(new DecelerateInterpolator(4));
@@ -99,8 +88,6 @@ public class LoginWithPhone extends AppCompatActivity {
         reenterSlide.setInterpolator(new DecelerateInterpolator(2));
         reenterSlide.addTarget(R.id.llphone);
         getWindow().setReenterTransition(reenterSlide);
-
-
     }
 
     Transition.TransitionListener enterTransitionListener = new Transition.TransitionListener() {
@@ -111,11 +98,9 @@ public class LoginWithPhone extends AppCompatActivity {
 
         @Override
         public void onTransitionEnd(Transition transition) {
-
             ivBack.setImageAlpha(255);
             Animation animation = AnimationUtils.loadAnimation(LoginWithPhone.this, R.anim.slide_right);
             ivBack.startAnimation(animation);
-
         }
 
         @Override
@@ -138,7 +123,6 @@ public class LoginWithPhone extends AppCompatActivity {
     Transition.TransitionListener returnTransitionListener = new Transition.TransitionListener() {
         @Override
         public void onTransitionStart(Transition transition) {
-
             InputMethodManager imm = (InputMethodManager) getSystemService(
                     Context.INPUT_METHOD_SERVICE);
             imm.hideSoftInputFromWindow(etPhoneNo.getWindowToken(), 0);
@@ -155,7 +139,6 @@ public class LoginWithPhone extends AppCompatActivity {
 
         @Override
         public void onTransitionEnd(Transition transition) {
-
 
         }
 
@@ -178,7 +161,6 @@ public class LoginWithPhone extends AppCompatActivity {
     Transition.TransitionListener exitTransitionListener = new Transition.TransitionListener() {
         @Override
         public void onTransitionStart(Transition transition) {
-
             rootFrame.setAlpha(1f);
             fabProgressCircle.hide();
             llPhone.setBackgroundColor(Color.parseColor("#EFEFEF"));
@@ -186,7 +168,6 @@ public class LoginWithPhone extends AppCompatActivity {
 
         @Override
         public void onTransitionEnd(Transition transition) {
-
 
         }
 
@@ -206,22 +187,18 @@ public class LoginWithPhone extends AppCompatActivity {
         }
     };
 
-
     Transition.TransitionListener reenterTransitionListener = new Transition.TransitionListener() {
         @Override
         public void onTransitionStart(Transition transition) {
-
 
         }
 
         @Override
         public void onTransitionEnd(Transition transition) {
-
             llPhone.setBackgroundColor(Color.parseColor("#FFFFFF"));
             etPhoneNo.setCursorVisible(true);
             InputMethodManager imm = (InputMethodManager) getSystemService(Context.INPUT_METHOD_SERVICE);
             imm.toggleSoftInput(InputMethodManager.SHOW_FORCED, 0);
-
         }
 
         @Override
@@ -243,22 +220,14 @@ public class LoginWithPhone extends AppCompatActivity {
 
     @OnClick(R.id.fabProgressCircle)
     void nextPager() {
-
         etPhoneNo.setCursorVisible(false);
         rootFrame.setAlpha(0.4f);
         fabProgressCircle.show();
 
-
-
-        new Handler().postDelayed(new Runnable() {
-            @Override
-            public void run() {
-
-                Intent intent = new Intent(LoginWithPhone.this, PasswordActivity.class);
-                ActivityOptionsCompat options = ActivityOptionsCompat.
-                        makeSceneTransitionAnimation(LoginWithPhone.this);
-                startActivity(intent, options.toBundle());
-            }
+        new Handler().postDelayed(() -> {
+            Intent intent = new Intent(LoginWithPhone.this, PasswordActivity.class);
+            ActivityOptionsCompat options = ActivityOptionsCompat.makeSceneTransitionAnimation(LoginWithPhone.this);
+            startActivity(intent, options.toBundle());
         }, 1000);
     }
 
