@@ -3,22 +3,14 @@ package mohak.uberux;
 import android.content.Context;
 import android.content.Intent;
 import android.os.Handler;
-import android.support.v4.app.ActivityOptionsCompat;
-import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.transition.Slide;
-import android.transition.Transition;
-import android.transition.TransitionInflater;
-import android.view.Gravity;
-import android.view.animation.AccelerateDecelerateInterpolator;
-import android.view.animation.AccelerateInterpolator;
-import android.view.animation.Animation;
-import android.view.animation.AnimationUtils;
 import android.view.animation.DecelerateInterpolator;
 import android.view.inputmethod.InputMethodManager;
 import android.widget.EditText;
 import android.widget.FrameLayout;
-import android.widget.Toast;
+
+import androidx.appcompat.app.AppCompatActivity;
 
 import com.github.jorgecastilloprz.FABProgressCircle;
 
@@ -26,13 +18,9 @@ import butterknife.BindView;
 import butterknife.ButterKnife;
 import butterknife.OnClick;
 
-import static android.view.Gravity.LEFT;
 import static android.view.Gravity.RIGHT;
 
-
 public class PasswordActivity extends AppCompatActivity {
-
-
     @BindView(R.id.fabProgressCircle)
     FABProgressCircle fabProgressCircle;
 
@@ -60,8 +48,6 @@ public class PasswordActivity extends AppCompatActivity {
         returnSlide.addTarget(R.id.llphone);
         returnSlide.setInterpolator(new DecelerateInterpolator());
         getWindow().setReturnTransition(returnSlide);
-
-
     }
 
 
@@ -75,14 +61,10 @@ public class PasswordActivity extends AppCompatActivity {
                 Context.INPUT_METHOD_SERVICE);
         imm.hideSoftInputFromWindow(etPass.getWindowToken(), 0);
 
-        new Handler().postDelayed(new Runnable() {
-            @Override
-            public void run() {
-
-                Intent intent = new Intent(PasswordActivity.this, MapActivity.class);
-                startActivity(intent);
-                finish();
-            }
+        new Handler().postDelayed(() -> {
+            Intent intent = new Intent(PasswordActivity.this, MapActivity.class);
+            startActivity(intent);
+            finish();
         }, 1000);
     }
 
