@@ -1,18 +1,13 @@
 package mohak.uberux;
 
-
 import com.google.gson.JsonObject;
 
 import retrofit2.Call;
 import retrofit2.http.GET;
 import retrofit2.http.Query;
 
-/**
- * Created by mohak on 2/8/17.
- */
-
 public interface getPolyline {
-
     @GET("json")
-    Call<JsonObject> getPolylineData(@Query("origin") String origin, @Query("destination") String destination);
+    Call<JsonObject> getPolylineData(@Query("origin") String origin, @Query("destination") String destination,
+            @Query("key") String key);
 }
